uplink_code/requirements.txt
anvil-uplink
python-dotenv
pyodbc